Extract vuex plugin necessary only for UI
