+++ src/store/index.js
@@ -10,18 +10,11 @@ import desktopModule from './modules/desktop';
 import mobileModule from './modules/mobile';
 import accountsModule from './modules/accounts';
 import persistState from './plugins/persistState';
-import ledgerConnection from './plugins/ledgerConnection';
 import remoteConnection from './plugins/remoteConnection';
-import notificationOnRemoteConnection from './plugins/notificationOnRemoteConnection';
-import desktopGuide from './plugins/desktopGuide';
 import initSdk from './plugins/initSdk';
-import modals from './plugins/modals';
 import registerServiceWorker from './plugins/registerServiceWorker';
-import browserPathTracker from './plugins/browserPathTracker';
-import observables from './plugins/observables';
 import reverseIframe from './plugins/reverseIframe';
 import syncLedgerAccounts from './plugins/syncLedgerAccounts';
-import connectionStatusTracker from './plugins/connectionStatusTracker';
 
 Vue.use(Vuex);
 Vue.use(VueRx);
@@ -81,14 +74,9 @@ const store = new Vuex.Store({
     ),
     initSdk,
     remoteConnection,
-    modals,
     registerServiceWorker,
-    observables,
     reverseIframe,
-    connectionStatusTracker,
-    ...process.env.IS_MOBILE_DEVICE
-      ? [notificationOnRemoteConnection, browserPathTracker]
-      : [ledgerConnection, syncLedgerAccounts, desktopGuide],
+    ...process.env.IS_MOBILE_DEVICE ? [] : [syncLedgerAccounts],
   ],
 
   modules: {
