+++ src/store/plugins/ui/__tests__/notificationOnRemoteConnection.js
