+++ src/store/plugins/ui/browserPathTracker.js
